+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.0.2] - 2026-01-21
+
+### Added
+- `check-updates` subcommand to explicitly check for available updates from PyPI
+- `upgrade` subcommand to upgrade midi-diff to the latest version using pip
+- `--pre` flag for `upgrade` subcommand to include pre-release versions
+
 ## [1.0.1] - 2026-01-21
 
 ### Changed